+++ resharper/AbpInsight.Rider.Tests/Analysis/PublishAbpEventAnalyzerTests.cs
@@ -1,0 +1,11 @@
+﻿using AbpInsight.Daemon.Stages.Highlightings;
+using NUnit.Framework;
+
+namespace AbpInsight.Rider.Tests.Analysis;
+
+[TestAbpInsight]
+public class PublishAbpEventAnalyzerTests : CSharpHighlightingTestBase<CSharpAbpInsightHighlightingBase>
+{
+    [Test]
+    public void PublishAbpEventAnalyzer() => DoNamedTest();
+}