+++ resharper/AbpInsight.Rider.Tests/test/data/Analysis/PublishAbpEventAnalyzer.cs
@@ -1,0 +1,10 @@
+﻿using Volo.Abp.EventBus;
+
+public class Test
+{
+    public void Test()
+    {
+        IEventBus eb = null;
+        eb.PublishAsync(new Test());
+    }
+}