+++ resharper/AbpInsight/Resources/Strings.Designer.cs
@@ -65,7 +65,8 @@ namespace AbpInsight.Resources
     public static string ModuleTypeMustBePublic_Message => ResourceManager.GetString("ModuleTypeMustBePublic_Message");
     public static string DependencyTypeMustBePublic_Message => ResourceManager.GetString("DependencyTypeMustBePublic_Message");
     public static string DependencyImplementsMultipleLifetimes_Message => ResourceManager.GetString("DependencyImplementsMultipleLifetimes_Message");
-    public static string DependsOnNonAbpModuleType_Message => ResourceManager.GetString("DependsOnNonAbpModuleType_Message");
     public static string AddAbpWidget_Description => ResourceManager.GetString("AddAbpWidget_Description");
+    public static string DependsOnNonAbpModuleType_Message => ResourceManager.GetString("DependsOnNonAbpModuleType_Message");
+    public static string PublishAbpEvent_Message => ResourceManager.GetString("PublishAbpEvent_Message");
   }
 }